Enhance UI components: update card styles, add TypographySection, improve button variants, and refine avatar and input elements
