+++ src/components/Design/Components.tsx
@@ -1,0 +1,210 @@
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
+import { Separator } from "@/components/ui/separator";
+import { Skeleton } from "@/components/ui/skeleton";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+
+import {
+  Mail,
+  Plus,
+  Search,
+  Settings,
+  User
+} from "lucide-react";
+
+const ComponentSection = () => {
+  return (
+    <>
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Buttons</h2>
+        <div className="flex flex-wrap gap-4">
+          <Button>Default</Button>
+          <Button variant="secondary">Secondary</Button>
+          <Button variant="success">Success</Button>
+          <Button variant="warning">Warning</Button>
+          <Button variant="destructive">Destructive</Button>
+          <Button variant="outline">Outline</Button>
+          <Button variant="ghost">Ghost</Button>
+          <Button variant="link">Link</Button>
+        </div>
+        <div className="flex flex-wrap gap-4">
+          <Button size="sm">Small</Button>
+          <Button>Default</Button>
+          <Button size="lg">Large</Button>
+          <Button size="icon">
+            <Settings />
+          </Button>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Inputs */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Inputs</h2>
+        <div className="grid gap-4 max-w-sm">
+          <Input placeholder="Default input" />
+          <Input placeholder="Outlined input" startIcon={Search} />
+          <Input placeholder="Disabled input" disabled />
+          <div className="flex gap-2">
+            <Input placeholder="With button" />
+            <Button>Send</Button>
+          </div>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Cards */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Cards</h2>
+        <div className="grid md:grid-cols-2 gap-6">
+          <Card>
+            <CardHeader>
+              <CardTitle>Card Title</CardTitle>
+              <CardDescription>Card Description</CardDescription>
+            </CardHeader>
+            <CardContent>
+              <p>Card content goes here.</p>
+            </CardContent>
+            <CardFooter>
+              <Button>Action</Button>
+            </CardFooter>
+          </Card>
+
+          <Card className="bg-blueprint">
+            <CardHeader>
+              <CardTitle>Card Title</CardTitle>
+              <CardDescription>Card Description</CardDescription>
+            </CardHeader>
+            <CardContent>
+              <p>Card content goes here.</p>
+            </CardContent>
+            <CardFooter>
+              <Button>Action</Button>
+            </CardFooter>
+          </Card>
+
+          <Card>
+            <CardHeader>
+              <div className="flex items-center gap-4">
+                <Avatar>
+                  <AvatarImage src="https://github.com/shadcn.png" />
+                  <AvatarFallback>CN</AvatarFallback>
+                </Avatar>
+                <div>
+                  <CardTitle>Profile Card</CardTitle>
+                  <CardDescription>With avatar</CardDescription>
+                </div>
+              </div>
+            </CardHeader>
+            <CardContent>
+              <p>Card with avatar example.</p>
+            </CardContent>
+          </Card>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Badges */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Badges</h2>
+        <div className="flex flex-wrap gap-4">
+          <Badge>Default</Badge>
+          <Badge variant="secondary">Secondary</Badge>
+          <Badge variant="success">Success</Badge>
+          <Badge variant="warning">Warning</Badge>
+          <Badge variant="destructive">Destructive</Badge>
+          <Badge variant="outline">Outline</Badge>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Avatar */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Avatars</h2>
+        <div className="flex flex-wrap gap-4">
+          <Avatar>
+            <AvatarImage src="https://github.com/shadcn.png" />
+            <AvatarFallback>CN</AvatarFallback>
+          </Avatar>
+          <Avatar>
+            <AvatarFallback>
+              <User className="w-4 h-4" />
+            </AvatarFallback>
+          </Avatar>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Tooltips */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Tooltips</h2>
+        <div className="flex flex-wrap gap-4">
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Button variant="outline" size="icon">
+                  <Plus className="h-4 w-4" />
+                </Button>
+              </TooltipTrigger>
+              <TooltipContent>
+                <p>Add new item</p>
+              </TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Button variant="outline" size="icon">
+                  <Mail className="h-4 w-4" />
+                </Button>
+              </TooltipTrigger>
+              <TooltipContent>
+                <p>Send email</p>
+              </TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        </div>
+      </div>
+
+      <Separator className="my-8" />
+
+      {/* Skeleton */}
+      <div className="space-y-4">
+        <h2 className="text-2xl font-semibold">Skeleton</h2>
+        <div className="space-y-4">
+          <div className="flex items-center space-x-4">
+            <Skeleton className="h-12 w-12 rounded-full" />
+            <div className="space-y-2">
+              <Skeleton className="h-4 w-[250px]" />
+              <Skeleton className="h-4 w-[200px]" />
+            </div>
+          </div>
+          <Skeleton className="h-32 w-full" />
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default ComponentSection;
