+++ src/components/ui/button.tsx
@@ -18,8 +18,8 @@ const buttonVariants = cva(
         success: "bg-success text-success-foreground shadow-sm hover:bg-success/90",
         warning: "bg-warning text-warning-foreground shadow-sm hover:bg-warning/90",
         destructive: "bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90",
-        outline: "border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground",
-        ghost: "hover:bg-accent hover:text-accent-foreground",
+        outline: "border border-input bg-background shadow-sm hover:border-foreground/70 hover:text-accent-foreground",
+        ghost: "hover:bg-foreground/10",
         icon: "",
         link: "text-primary underline-offset-4 hover:underline",
       },
