+++ src/index.css
@@ -38,16 +38,18 @@
   :root {
     --background: 220 13% 98%;
     --foreground: 220 13% 20%;
+    --foreground-muted: 220 13% 30%;
 
-    --blueprint: 238, 93%, 84%;
-    --blueprint-foreground: 220 13% 45;
+    --blueprint: 238 93% 84%;
+    --blueprint-foreground: 220 13% 45%;
+    --blueprint-foreground-muted: 220 13% 45%;
 
-    --extended: 242 91% 75%;
-    --extended-foreground: 51 33% 88%;
-    
     --surface-1: 220 13% 90%;
     --surface-2: 220 13% 80%;
     --surface-3: 220 13% 70%;
+
+    --semantic-foreground: 0 0% 90%;
+    --semantic-foreground-muted: 0 0% 80%;
 
     --card: var(--surface-1);
     --card-foreground: var(--foreground);
@@ -57,29 +59,29 @@
     --popover-foreground: 220 13% 3.9%;
  
     --primary: 242 91% 75%;
-    --primary-foreground: 0 0% 100% / 0.9;
+    --primary-foreground: var(--semantic-foreground);
  
-    --secondary: 265 44% 32%;
-    --secondary-foreground: 0 0% 100% / 0.9;
+    --secondary: 242 44% 59%;
+    --secondary-foreground: var(--semantic-foreground);
 
     --success: 90 36% 38%;
-    --success-foreground: 0 0% 100% / 0.9;
+    --success-foreground: var(--semantic-foreground);
 
     --warning: 49 100% 38%;
-    --warning-foreground: 0 0% 100% / 0.9;
+    --warning-foreground: var(--semantic-foreground);
  
     --destructive: 3 57% 36%;
-    --destructive-foreground: 0 0% 100% / 0.9;
+    --destructive-foreground: var(--semantic-foreground);
     
-    --accent: 0 0% 96.1%;
+    --accent: 265 44% 32%;
     --accent-foreground: 0 0% 9%;
     
-    --muted: 220 13% 95%;
-    --muted-foreground: 220 13% 45;
+    --muted: var(--surface-3);
+    --muted-foreground: 220 13% 95%;
 
-    --border: 220 13% 91%;
-    --input: 220 13% 91%;
-    --ring: 220 13% 3.9%;
+    --border: var(--surface-3);
+    --input: var(--surface-2);
+    --ring: var(--primary);
  
     --radius: 0.5rem;
  
@@ -94,18 +96,20 @@
   }
  
   .dark {
-    --background: 0, 0%, 10%;
+    --background: 0 0% 10%;
     --foreground: 0 0% 98%;
+    --foreground-muted: 0 0% 85%;
   
-    --blueprint: 239, 93%, 84%;
-    --blueprint-foreground: 220 13% 45;
+    --blueprint: 239 36% 67%;
+    --blueprint-foreground: 220 5% 90%;
+    --blueprint-foreground-muted: 220 5% 80%;
   
-    --extended: 242 91% 75%;
-    --extended-foreground: 51 33% 88%;
-    
     --surface-1: 0 0% 17%;
     --surface-2: 0 0% 25%;
     --surface-3: 0 0% 32%;
+
+    --semantic-foreground: 0 0% 90%;
+    --semantic-foreground-muted: 0 0% 80%;
   
     --card: var(--surface-1);
     --card-foreground: var(--foreground);
@@ -114,30 +118,30 @@
     --popover: var(--surface-1);
     --popover-foreground: var(--foreground);
   
-    --primary: var(--extended);
-    --primary-foreground: 0 0% 100% / 0.9;
+    --primary: 242 91% 75%;
+    --primary-foreground: var(--surface-foreground);
   
-    --secondary: 265, 44%, 32%;
-    --secondary-foreground: 0 0% 100% / 0.9;
+    --secondary: 246 100% 77%;
+    --secondary-foreground: var(--surface-foreground);
+    
+    --accent: 265 44% 32%;
+    --accent-foreground: 0 0% 98%;
   
     --success: 90 36% 38%;
-    --success-foreground: 0 0% 100% / 0.9;
+    --success-foreground: var(--surface-foreground);
   
     --warning: 49 100% 38%;
-    --warning-foreground: 0 0% 100% / 0.9;
+    --warning-foreground: var(--surface-foreground);
   
     --destructive: 3 57% 36%;
-    --destructive-foreground: 0 0% 100% / 0.9;
-    
-    --accent: 0 0% 14.9%;
-    --accent-foreground: 0 0% 98%;
+    --destructive-foreground: var(--surface-foreground);
   
     --muted: var(--surface-3);
     --muted-foreground: 0 0% 70%;
   
     --border: var(--surface-3);
     --input: var(--surface-2);
-    --ring: 213 36% 50%;
+    --ring: var(--primary);
   
     --sidebar-background: var(--surface-1);
     --sidebar-foreground: var(--foreground);
@@ -151,37 +155,14 @@
 }
  
 @layer base {
-  * {
-    @apply border-border;
-  }
-  
-  body {
-    @apply bg-background text-foreground;
-  }
-
-  h1 {
-    @apply text-4xl font-bold font-minecraft mb-8;
-  }
-
-  h2 {
-    @apply text-2xl font-semibold font-minecraft mb-4;
-  }
-
-  h3 {
-    @apply text-xl font-semibold font-minecraft mb-3;
-  }
-
-  p {
-    @apply mb-4 text-muted-foreground;
-  }
-
-  a {
-    @apply text-primary hover:text-primary/90 transition-colors;
-  }
-
-  section {
-    @apply py-8;
-  }
+  * { @apply border-border; }
+  body { @apply bg-background text-foreground; }
+  h1 { @apply text-4xl font-bold font-minecraft mb-8; }
+  h2 { @apply text-2xl font-semibold font-minecraft mb-4; }
+  h3 { @apply text-xl font-semibold font-minecraft mb-3; }
+  p { @apply mb-4; }
+  a { @apply text-primary hover:text-primary/90 transition-colors; }
+  section { @apply py-8; }
 }
 
 .bg-andesite-casing {
